# proto/friend.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "proto/friend.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "FriendService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/add-friend": {
      "post": {
        "operationId": "FriendService_AddFriend",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/friendAddFriendResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/friendAddFriendRequest"
            }
          }
        ],
        "tags": [
          "FriendService"
        ]
      }
    },
    "/mutual-friends": {
      "get": {
        "operationId": "FriendService_GetMutualFriends",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/friendMutualFriendsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user1",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "user2",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "tags": [
          "FriendService"
        ]
      }
    },
    "/mutual-friends/{user1}/{user2}": {
      "get": {
        "operationId": "FriendService_GetMutualFriends2",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/friendMutualFriendsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "user1",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "user2",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "FriendService"
        ]
      }
    }
  },
  "definitions": {
    "friendAddFriendRequest": {
      "type": "object",
      "properties": {
        "from": {
          "type": "string"
        },
        "to": {
          "type": "string"
        }
      }
    },
    "friendAddFriendResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    },
    "friendMutualFriendsResponse": {
      "type": "object",
      "properties": {
        "mutualFriends": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}